you can shoot targets now

+++ src/Game.h
@@ -5,7 +5,7 @@ typedef struct Target_t
 {
     bool isActive;
     int posX, posY;
-    double despawnTime;
+    unsigned long despawnTime;
 } Target_t;
 
 #endif //GAME_H
